+++ chic-mbg/src/main/java/com/github/chic/entity/RolePermissionRelation.java
@@ -20,15 +20,15 @@ public class RolePermissionRelation implements Serializable {
      * ID
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 角色ID
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 权限ID
      */
-    private Integer permissionId;
+    private Long permissionId;
 }
